+++ tests/test_comparisons.py
@@ -219,6 +219,15 @@ class TestSameMeaning:
 
 
 class TestImplication:
+    def test_comparison_implies_predicate_false(self):
+        distance = Comparison(
+            "the distance between $place1 and $place2 was",
+            sign=">",
+            expression="20 miles",
+        )
+        predicate = Predicate("the distance between $place1 and $place2 was")
+        assert not distance.implies(predicate)
+
     def test_predicate_not_same_with_interchangeable_terms(self):
         interchangeable = Comparison(
             "the distance between $place1 and $place2 was",